refactor: main js to be auth js
