+++ index.html
@@ -306,7 +306,6 @@
       </div>
     </footer>
 
-    <script type="module" src="src/js/main.js"></script>
-    <script type="module" src="src/js/logout.js"></script>
+    <script type="module" src="src/js/auth.js"></script>
   </body>
 </html>
